Exchange complex type back and forth with mapping
This passes a complex type `Vector3f` back and forth between the host and the plugin, but the "host -> plugin" and "plugin -> host" directions use different types!

This forces the plugin to convert from one type to the other before calling the host function. And similar problems could come up on host in a more realistic example.